backend/books/views.py
from rest_framework import filters, viewsets

from .models import Book
from .serializers import BookSerializer


class BookViewSet(viewsets.ModelViewSet):
    queryset = Book.objects.all()
    serializer_class = BookSerializer
    filter_backends = [filters.SearchFilter, filters.OrderingFilter]
    search_fields = ["title", "author", "category"]
    ordering_fields = ["title", "price", "created_at"]
    ordering = ["title"]
