backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.1.1
